requirements.txt
bcrypt==3.2.0
botocore==1.22.2
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
fabric==2.6.0
invoke==1.6.0
jmespath==0.10.0
paramiko==2.8.0
pathlib2==2.3.6
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
